refactor(env): share dotenv serializer so `export` escapes like `run --export`
Move quoteDotenvValue/toDotenv into src/utils/dotenv.js and use it from both
the CLI runner and EnvironmentVaultService.exportEnv. Previously `vault env
export --format dotenv` wrote raw key=value pairs, so values with newlines,
quotes, # or edge whitespace produced output that did not round-trip through
`vault env import`. Now both paths quote/escape identically.

envRunHelpers re-exports the helpers for existing importers. Closes the
follow-up noted in SPEC §16.10.
